Updated Md details table to row component along with edit and delete confirmation.
